photo-classify-tg-bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.1
aiohttp_socks
imagehash==4.3.1
Pillow==9.5.0
pyrogram==2.0.104
tgcrypto
